pyproject.toml
[tool.poetry]
name = "NBSDynamics"
version = "0.3.0"
description = "A biophysical model framework written in Python to make simulations on coral development based on four environmental factors: (1) light; (2) flow; (3) temperature; and (4) acidity. For the hydrodynamics, the model can be coupled to Delft3D Flexible Mesh; a hydrodynamic model developed at Deltares."
authors = ["Jasper Dijkstra <jasper.dijkstra@deltares.nl>"]
maintainers = ["Carles S. Soriano Perez <carles.sorianoperez@deltares.nl>"]
license = "MIT"
packages = [
  { include = "src" },
]

[tool.poetry.dependencies]
python = "^3.8"
netCDF4 = "^1.4.2"
numpy = "^1.14.3"
scipy = "^1.1.0"
matplotlib = "^2.2.2"
pandas = "^1.3.4"
tqdm = "^4.62.3"
isort = "^5.9.3"
black = "^21.9b0"
bmi-python = "^0.3.0"
pydantic = "^1.8.2"
pypiwin32 = "^223"
pywin32 = "^302"

[tool.poetry.dev-dependencies]
pytest = "^6.2"
black = { version = "*", allow-prereleases = true }
mkdocs = "^1.1"
mkdocs-material = "^7.1"
mkdocstrings = "^0.15"
mkdocs-autorefs = "^0.1"
pytest-cov = "^2.11"
pymdown-extensions = "^8.1"
flake8 = "^3.9.2"
mypy = "^0.910"
devtools = "^0.6.1"
commitizen = "^2.20.0"
pre-commit = "^2.15.0"

[tool.commitizen]
name = "cz_conventional_commits"
changelog_file = "docs/changelog.md"
update_changelog_on_bump = true
version = "0.3.0"
tag_format = "v$major.$minor.$patch"
version_files = [
    "src/core/__init__.py",
    "pyproject.toml:version"
]

[tool.commitizen.custommize]
bump_pattern = "^(break|new|fix|hotfix|refactor|docs)"
bump_map = {"break" = "MAJOR", "new" = "MINOR", "fix" = "PATCH", "hotfix" = "PATCH", "refactor"="PATCH", "docs" = "PATCH"}

[tool.coverage.run]
omit = [
    "test/*",
    "docs/*"
    ]

[tool.black]
line-length = 88
target-version = ['py38', 'py39']
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | \.virtualenvs
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
